UBUNTU: [Config] perarch and indep tools builds need separate build directories
Backported from a Quantal patch by Andy Whitcroft <apw@canonical.com>

We are losing the perarch perf binary etc when building, this is occuring
based on parallel execution order.  This occurs because both build passes
use 'debian/build/tools' as their build directory.  This is simply not
safe as execution order is not guarenteed.  Move the perarch build to
tools-perarch to avoid collisions.

Signed-off-by: Tim Gardner <tim.gardner@canonical.com>